requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.15.0
sqlalchemy==2.0.19
openai==1.3.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
numpy==1.24.3
